agents/care-management/pre-call-assessment/services/agent-runtime/pyproject.toml
[tool.poetry]
name = "agent-runtime"
version = "0.1.0"
description = "LangGraph agent runtime"
authors = ["you"]
package-mode = false

[tool.poetry.dependencies]
requests = "^2.32.0"
python = "^3.11"
langgraph = { version = "^1.0.8", extras = ["sqlite"] }
pyyaml = "^6.0"
python-dotenv = "^1.0.0"  # For managing environment variables
fastapi = "^0.129.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
boto3 = "^1.42.0"
langchain = "^1.2.10"
langchain-openai = "^1.1.10"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
jsonschema = "^4.22.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.4.0"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]



